requirements.txt
meltano==0.2.1
pymongo==3.11.2
python-dotenv==0.15.0